Test part 2

+++ GliToJiraImporter.Testing/Tests/ParserTester.cs
@@ -23,8 +23,8 @@ namespace GliToJiraImporter.Testing.Tests
     {
         private static readonly ILog log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
 
-        private static readonly string checkoffPath = @"Public\TestCheckoffs\";
-        private static readonly string expectedResultPath = @"Public\ExpectedResults\";
+        private static readonly string checkoffFolderName = @"Public\TestCheckoffs\";
+        private static readonly string expectedResultFolderName = @"Public\ExpectedResults\";
         private const string CLAUSE_ID = "customfield_10046";
         private const string CATEGORY = "customfield_10044";
         private const string SUBCATEGORY = "customfield_10045";
@@ -65,7 +65,7 @@ namespace GliToJiraImporter.Testing.Tests
 
             parameterModelStub = new()
             {
-                FilePath = $"{checkoffPath}Australia-New-Zealand.docx",
+                FilePath = $"{checkoffFolderName}Australia-New-Zealand.docx",
                 Method = new HttpMethod("GET"),
                 JiraUrl = "https://gre-team.atlassian.net/rest/api/2",//search?jql=project=EGRE&maxResults=10",
                 UserName = userNameToken,
@@ -122,25 +122,29 @@ namespace GliToJiraImporter.Testing.Tests
             //IList<string> categories = expectedResult.Select(cat => cat.Category).ToList();
             //Dictionary<string, Issue> issues = (Dictionary<string, Issue>)this.jiraConnectionStub.Issues.GetIssuesAsync().Result;
             //foreach (Models.Issue issue in jiraExistingIssueList)
-            for (int i = 0; i < this.expectedResult.Count; i++)
+            if (jiraExistingIssueList.Count > 0)
             {
-                for (int j = 0; j < this.expectedResult[i].RegulationList.Count; j++)
+
+                for (int i = 0; i < this.expectedResult.Count; i++)
                 {
-                    //Models.Issue issue = jiraExistingIssueList[i];
-                    Models.Issue issueFound = jiraExistingIssueList.First(issue => issue.fields.customfield_10046.Equals(this.expectedResult[i].RegulationList[j].ClauseID.FullClauseId));
-                    if (issueFound.fields.customfield_10046 != null && issueFound.fields.customfield_10046.Equals(this.expectedResult[i].RegulationList[j].ClauseID.FullClauseId))//categories.Contains(issue["GLICategory"].Value) && issue.Labels.Count() == 0)//TODO not a good enough check
+                    for (int j = 0; j < this.expectedResult[i].RegulationList.Count; j++)
                     {
-                        bool success = jiraRequestUtilities.DeleteIssueByKey(issueFound.key);//this.deleteIssueByKey(issue.Key.Value);
-                        if (success != true)
+                        //Models.Issue issue = jiraExistingIssueList[i];
+                        Models.Issue issueFound = jiraExistingIssueList.First(issue => issue.fields.customfield_10046.Equals(this.expectedResult[i].RegulationList[j].ClauseID.FullClauseId));
+                        if (issueFound.fields.customfield_10046 != null && issueFound.fields.customfield_10046.Equals(this.expectedResult[i].RegulationList[j].ClauseID.FullClauseId))//categories.Contains(issue["GLICategory"].Value) && issue.Labels.Count() == 0)//TODO not a good enough check
                         {
-                            log.Error($"Issue failed to delete. {issueFound.key}");
+                            bool success = jiraRequestUtilities.DeleteIssueByKey(issueFound.key);//this.deleteIssueByKey(issue.Key.Value);
+                            if (success != true)
+                            {
+                                log.Error($"Issue failed to delete. {issueFound.key}");
+                            }
                         }
+                        Thread.Sleep(this.parameterModelStub.SleepTime);
                     }
-                    Thread.Sleep(this.parameterModelStub.SleepTime);
+                    expectedResult.Clear();
                 }
-                expectedResult.Clear();
+                checkForErrorsInLogs();
             }
-            checkForErrorsInLogs();
         }
 
         //[Ignore("Can only run locally with a local Jira.")]
@@ -148,8 +152,8 @@ namespace GliToJiraImporter.Testing.Tests
         public void ParserSingleTest()
         {
             //given
-            parameterModelStub.FilePath = $"{checkoffPath}SINGLE-Australia-New-Zealand.docx";
-            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{expectedResultPath}ParserSingleTestExpectedResult.json"));
+            parameterModelStub.FilePath = $"{GetFolderPath(checkoffFolderName)}SINGLE-Australia-New-Zealand.docx";
+            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{GetFolderPath(expectedResultFolderName)}ParserSingleTestExpectedResult.json"));
 
             //when
             IList<CategoryModel> result = sut.Parse();
@@ -163,8 +167,8 @@ namespace GliToJiraImporter.Testing.Tests
         public void ParserSingleMultiDescTest()
         {
             //given
-            parameterModelStub.FilePath = $"{checkoffPath}SINGLE-MULTIDESC-Australia-New-Zealand.docx";
-            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{expectedResultPath}ParserSingleMultiDescTestExpectedResult.json"));
+            parameterModelStub.FilePath = $"{GetFolderPath(checkoffFolderName)}SINGLE-MULTIDESC-Australia-New-Zealand.docx";
+            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{GetFolderPath(expectedResultFolderName)}ParserSingleMultiDescTestExpectedResult.json"));
 
             //when
             IList<CategoryModel> result = sut.Parse();
@@ -178,8 +182,8 @@ namespace GliToJiraImporter.Testing.Tests
         public void ParserPicturesTest()
         {
             //given
-            parameterModelStub.FilePath = $"{checkoffPath}PICTURES-SHORT-Australia-New-Zealand.docx";
-            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{expectedResultPath}PicturesTestExpectedResult.json"));
+            parameterModelStub.FilePath = $"{GetFolderPath(checkoffFolderName)}PICTURES-SHORT-Australia-New-Zealand.docx";
+            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{GetFolderPath(expectedResultFolderName)}PicturesTestExpectedResult.json"));
 
             //when
             IList<CategoryModel> result = sut.Parse();
@@ -194,8 +198,8 @@ namespace GliToJiraImporter.Testing.Tests
         public void ParserSpecialsTest()
         {
             //given
-            parameterModelStub.FilePath = $"{checkoffPath}SPECIALS-Australia-New-Zealand.docx";
-            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{expectedResultPath}ParserSpecialsTestExpectedResult.json"));
+            parameterModelStub.FilePath = $"{GetFolderPath(checkoffFolderName)}SPECIALS-Australia-New-Zealand.docx";
+            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{GetFolderPath(expectedResultFolderName)}ParserSpecialsTestExpectedResult.json"));
 
             //when
             IList<CategoryModel> result = sut.Parse();
@@ -210,8 +214,8 @@ namespace GliToJiraImporter.Testing.Tests
         public void ParserFullTest()
         {
             //given
-            //expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{expectedResultPath}FullTestSearchExpectedResult.json"));
-            parameterModelStub.FilePath = $"{checkoffPath}Australia-New-Zealand.docx";
+            //expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{GetFolderPath(expectedResultFolderName)}FullTestSearchExpectedResult.json"));
+            parameterModelStub.FilePath = $"{GetFolderPath(checkoffFolderName)}Australia-New-Zealand.docx";
             int expectedCount = 633;
             //when
             IList<CategoryModel> result = sut.Parse();
@@ -232,7 +236,7 @@ namespace GliToJiraImporter.Testing.Tests
         public void ParserUnknownDocTypeTest()
         {
             //given
-            parameterModelStub.FilePath = $"{checkoffPath}SINGLE-Australia-New-Zealand.docx";
+            parameterModelStub.FilePath = $"{GetFolderPath(checkoffFolderName)}SINGLE-Australia-New-Zealand.docx";
             parameterModelStub.Type = 0;
             IList<CategoryModel> result = new List<CategoryModel>();
             //when
@@ -253,8 +257,8 @@ namespace GliToJiraImporter.Testing.Tests
         public void ParserCharFormatTest()
         {
             //given
-            parameterModelStub.FilePath = $"{checkoffPath}CHARFORMAT-Australia-New-Zealand.docx";
-            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{expectedResultPath}ParserCharFormatTestExpectedResult.json"));
+            parameterModelStub.FilePath = $"{GetFolderPath(checkoffFolderName)}CHARFORMAT-Australia-New-Zealand.docx";
+            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{GetFolderPath(expectedResultFolderName)}ParserCharFormatTestExpectedResult.json"));
 
             //when
             IList<CategoryModel> result = sut.Parse();
@@ -269,8 +273,8 @@ namespace GliToJiraImporter.Testing.Tests
         public void ParserClauseIdVarietiesTest()
         {
             //given
-            parameterModelStub.FilePath = $"{checkoffPath}CLAUSEID-VARIETIES.docx";
-            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{expectedResultPath}ParserClauseIdVarietiesTestExpectedResult.json"));
+            parameterModelStub.FilePath = $"{GetFolderPath(checkoffFolderName)}CLAUSEID-VARIETIES.docx";
+            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{GetFolderPath(expectedResultFolderName)}ParserClauseIdVarietiesTestExpectedResult.json"));
 
             //when
             IList<CategoryModel> result = sut.Parse();
@@ -285,8 +289,8 @@ namespace GliToJiraImporter.Testing.Tests
         public void ParserNoCategoryTest()
         {
             //given
-            parameterModelStub.FilePath = $"{checkoffPath}NO-CATEGORY-Australia-New-Zealand.docx";
-            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{expectedResultPath}ParserNoCategoryTestExpectedResult.json"));
+            parameterModelStub.FilePath = $"{GetFolderPath(checkoffFolderName)}NO-CATEGORY-Australia-New-Zealand.docx";
+            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{GetFolderPath(expectedResultFolderName)}ParserNoCategoryTestExpectedResult.json"));
 
             //when
             IList<CategoryModel> result = sut.Parse();
@@ -301,8 +305,8 @@ namespace GliToJiraImporter.Testing.Tests
         public void ParserLinkTest()
         {
             //given
-            parameterModelStub.FilePath = $"{checkoffPath}LINKs-Australia-New-Zealand.docx";
-            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{expectedResultPath}ParserLinkTestExpectedResult.json"));
+            parameterModelStub.FilePath = $"{GetFolderPath(checkoffFolderName)}LINKs-Australia-New-Zealand.docx";
+            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{GetFolderPath(expectedResultFolderName)}ParserLinkTestExpectedResult.json"));
 
             //when
             IList<CategoryModel> result = sut.Parse();
@@ -317,8 +321,8 @@ namespace GliToJiraImporter.Testing.Tests
         public void ParserSingleDuplicateTest()
         {
             //given
-            parameterModelStub.FilePath = $"{checkoffPath}SINGLE-Australia-New-Zealand.docx";
-            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{expectedResultPath}ParserSingleTestExpectedResult.json"));
+            parameterModelStub.FilePath = $"{GetFolderPath(checkoffFolderName)}SINGLE-Australia-New-Zealand.docx";
+            expectedResult = JsonSerializer.Deserialize<List<CategoryModel>>(File.ReadAllText($"{GetFolderPath(expectedResultFolderName)}ParserSingleTestExpectedResult.json"));
 
             //when
             IList<CategoryModel> result = sut.Parse();
@@ -388,5 +392,12 @@ namespace GliToJiraImporter.Testing.Tests
 
         //    return success;
         //}
+
+        private string GetFolderPath(string folderName)
+        {
+            var currentAssemblyPath = Path.GetDirectoryName(Assembly.GetExecutingAssembly().CodeBase).Replace(@"file:\", string.Empty);
+            var relativePath = Path.Combine(currentAssemblyPath, @"..\..\..\" + folderName);
+            return Path.GetFullPath(relativePath);
+        }
     }
 }
